Allow the creation of a dynamic stub on the client end for the proxy.
The previous code had the proxy being forwarded over the wire to the
actual client, which meant that the client needs the handler class
definition. Obviate the need for this by exporting the proxy as a
remote object via UnicastRemoteObject.exportObject(remote,port).

This commit also refactors things a bit -- we don't need the class
name of the target stub, as it can be derived from getting the stub
itself.