Separate arg saving from return value/throwable saving in call record.
Previously, the call record was created to include the return
value. This might be too late to save the arguments (as they might
have been changed by the call to the method). So, save the original
args at call record creation, then save the return value or throwable
after.

To further insulate the call record, serialize the object arguments
into a byte array, and customize the serialization/deserialization of
the call record to use the byte array instead of the argument array.