Allow configuration of RMI proxy.
Particularly, allow the configuration of the proxy naming port, the
proxy instance port (as binding of a dynamic proxy only works when
using UnicastRemoteObject.exportObject(Remote, int) and not
UnicastRemoteObject.exportObject(Remote)), and target RMI name.

Also, allow the method recorder to be configured, with the default
being a do-nothing method recorder.