Support overloaded methods.
Unfortunately, the existing code does not support overloaded method
names, as it ties the method name directly to its arguments. So, foo()
and foo(bar) will both be mapped to the same set of arguments, and the
following sequence of calls:

   foo(bar)
   ...
   foo()

will fail on the second foo, as we've already recorded that foo has
one argument; conversely, if foo() was the first call, the above might
work with foo(bar) as we're not expecting any arguments with foo.

So, to remedy this, we add type information to the method name (i.e.,
we do a form of name mangling), and this new method name is then used
in mapping method names and their associated arguments.