Refactored scriptlet generation into separate util class.
The scriptlet generator is a singleton that knows how to represent
primitives, unpack arrays, and does introspection/reflection on
user-defined classes for their properties/public fields.

Currently, the generator uses stupid names for variables, and doesn't
unpack collections (yet). It should be fairly simple to unpack
collections, and I'm working on a way to create meaningful variable
names.