Make BeanShell interpreter for arguments a temporary.
Since JMeter serializes (into XML) all properties set via
setProperty*(), we can't have the BeanShell interpreter serialized;
trying to serialize the Interpreter instance would cause a
StackOverflowError as the Interpreter maintains a whole lot of state.

To deal with this, just mark the Interpreter as temporary, and
initialize the argument script when we set it up.