Recreate arguments after packing into a byte array.
Since custom serialization code of each argument is outside of our
control, and since such serialization could potentially change the
state of the argument while serializing it (which is a Bad Thing), we
need each argument as it was before the serialization. We assume that
deserializing each argument from the packed arguments byte array will
restore them to the state they were before serialization.

Therefore, unpack the arguments and use the freshly deserialized
objects as the arguments, discarding the ones we used for packing.