Allow exceptions to be ignored when sampling RMI.
Allow RMI invocation to throw exceptions, which are then ignored as
non-failing. Normally, if an RMI method is invoked and subsequently
throws an exception, it is treated as a failure. In contrast, some
methods may throw exceptions and the thrown exception is expected. In
this case, the user should check this box and have a BeanShell
assertion to check that an exception has been thrown.