Add support for building remote proxies via CGLIB.
We need to proxy non-root RMI stubs (i.e. classes extending
UnicastRemoteObject, implementing java.rmi.Remote, and which are part
of the object graph of a remote method return value); we cannot simply
replace instances of java.rmi.Remote with JDK proxies, since those
instances will not be treated specially by the JVM RMI infrastructure,
since those proxies do not extend UnicastRemoteObject -- what ends up
happening is that the JMeter JVM tries to serialize the whole proxy to
the actual client, instead of the JVM instructing the client JVM to
generate a stub.

Because of the need to have a proxy instance whose class extends
UnicastRemoteObject, we need to resort to using CGLIB to generate such
a proxy for us, with the caveat that we need to do the call handling
in a weird way. Because a) JDK proxies being final, and we cannot
extend final classes; and b) we need to extend UnicastRemoteObject,
and the methods that the real client is invoking do not exist on
parent -- we need to bypass the proxy instance and invoke those
methods on our stub instance through MethodProxy.

We retain the use of JDK proxies for root RMI remotes, since in those
cases using the CGLIB proxy has the opposite problem -- you can't
register a UnicastRemoteObject.

We also move the proxy building into DynamicStubProxyInvocationHandler
so as to DRY the code, as we will need to build proxies for non-root
RMI stubs later on.