Change log level on RMI sampler log messages.
This reduces the sample time as much as possible, and reduces the
chattiness in the JMeter log. We also flag EvalError as a warning or
error in the log, as appropriate -- we flag evaluation errors during
initial eval of the arguments script as a warning, as those are
potentially non-fatal to the run, while we flag calls to get the
current method arguments as an error.

NB: We may want to trigger a sampler failure on method argument script
evaluation errors later on, to make it more visible that there's a
problem constructing the arguments for the RMI method.