RMI Sampler: Measure eval and setup time as latency.
Although technically incorrect, we measure the time it takes to set up
the reference to the remote and the evaluation of the method arguments
script as latency, to make it more transparent in the results,
particularly so as to make it clear how much overhead BeanShell gives
us for each method call.

In particular, it is very difficult to measure actual network
latency (from TCP connection, network send, and first byte), as the
RMI spec demands that a socket factory is serialized from the RMI
server end for use by the client -- this means that in practice, we
cannot replace the socket factory on our end transparently, as the RMI
server we are connecting to may be providing its own client socket
factory (e.g. for encrypting the stream via SSL). There doesn't seem
to be any way to override or wrap the given client socket factory.

So, let's just reuse the latency measurement to measure our overhead
instead.

Design note: the existing RMI Sampler resuses the same remote instance
for all calls, potentially reusing the same socket for all
connections, acting as a bottleneck, and therefore may not be
representative of actual performance.