Properly bind registry to thread.
We were doing the wrong thing, misunderstanding how JMeter's test
variables are scoped.

When bound during test start, a variable is apparently scoped to the
whole of the test and is shared by all threads in the thread group;
doing the .putObject() at test start creates a single Registry for all
threads. I'm not particularly sure how exactly this happens (I have to
consult the JMeter source), but we do end up with the registry shared
across virtual users -- which is wrong, as we want each virtual
user/thread to have a different set of remote instances, including
root instances.

So, we change RMIRemoteObjectConfig to listen on thread start to
properly bind the registry to each thread context, making sure that
each thread has its own copy of the registry and therefore its own
remote instances.