RMI Sampler: Make BeanShell interpreter for args thread-local.
Make the BeanShell interpreter instance be a thread local, and don't
depend on the JMeter environment for holding it. This also means that
we don't have to worry (or understand) the concurrency semantics of
JMeter's context/variables/etc., and we can ensure that calling `.set`
on an interpreter instance will only affect that instance -- we don't
have to synchronize around the call to methodArgs() in the argument
script.