Tag method argument scriptlet with an ID.
This should help when debugging errors in the BeanShell
scriptlet. Since we log the evaluation error's message, and since the
message contains an ellipsized version of the scriptlet, the first
line of the scriptlet is useful for determining which scriptlet of
which sampler failed. Having a comment on top with a unique ID helps
in searching for that scriptlet.