Move argument packing functions to static helpers.
Move the argument packing and unpacking functions in MethodCallRecord
to static functions that we can import later in the RMI Sampler. We'll
use these to safely pack/unpack arguments to the call. Hopefully, the
overhead of serialization/deserialization isn't too much.

NB: It's actually better to use a custom RMI socket factory so that we
get actual on-the-wire sizes; using an ObjectOutputStream misses some
bytes from the RMI overhead (i.e. JRMP protocol overhead, etc.).