Pack/unpack args in sampler to measure size of sent args.
Pack and unpack args before invoking remote method, so we can measure
the size of the arguments as serialized bytes. We pack then unpack the
arguments, because we cannot trust that the serialized objects don't
have custom serialization logic with state-changing logic. However,
this may still fail due to deserializing an argument with custom logic
depending on global state -- we fall back to the existing args in that
case.