Introduce method for generating variable names/name parts.
Introduce a method for generating variable names, or parts of variable
names, for use in scriptlets where the variable name contains the type
of the variable.

As we either suffix each variable with a digit (for array/list element
numbers, or argument positions), or prefix with a path, disambiguating
between vars of the same type should be fine for the most part.