Add Objenesis as dependency.
We'll use this to construct classes which don't have default
constructors, particularly when they are serializable. As
serialization doesn't demand the presence of a default constructor,
some serializable classes being passed in as arguments may not have
default constructors so are impossible to create instances of, hence
our use of Objenesis.