Add helper method for constructing class instances via Objenesis.
In particular, this method will be used in scriptlets to generate
instances for classes which are serializable and do not declare a
default constructor. We use the RMIRemoteObjectConfig for containing
the Objenesis instance, so that we reconstruct the Objenesis cache
every time; otherwise, we may get errors when a different test gets
run, if we keep it in a static context.