Choose how to construct instances in generated scriptlet.
When generating scriptlets for arguments, opt to choose how to
construct object instances. If the object's class does not declare a
default constructor, inject the use of our Objenesis helper to create
an instance of the class; otherwise go through the default
constructor.

In this way, we avoid some of the possible construction/reflection
overhead due to Objenesis for classes that don't need to be
constructed in that way.