In proxy, give sampler a "descriptive" name.
When the proxy generates samplers, it should give each a fairly
descriptive name that distinguishes each individual element. This
allows us to properly disambiguate samplers in the log when e.g. an
exception is thrown. We include the target name as part of the sampler
name to distinguish between the methods with the same name on two
different remote objects. The call index is also included to
distinguish between series of calls to the same method on the same
remote.