Include remote target in sample label.
Distinguish between methods having the same name on different remote
objects in the results by including the target name in the sample
label. This means that a call to foo() on the root remote will be
grouped as a different sample to a call to foo() on a remote returned
by the root remote.