Don't generate scriptlets for non-Serializable instances.
As these objects should never be transmitted across the wire, leave
them out. Do note that this may fail in some situations -- if the
system being tested has serialization logic that serializes these
values in the containing object (e.g. class A is not serializable;
class B is Externalizable and manually reads & writes class A values).