Proxy: Use method name and object path for remote instance name.
When building the instance name for use in the generated RMI sampler,
use the method name (including call index) and the path to the remote
object from the relevant method's return value. This makes the value
of the target relatively stable across test recordings, instead of the
previous use of the remote's hash code.

Note that multiple recordings may still fail if the call index doesn't
match: if a different recording has a slightly different index for a
given method call, then the mapped remote name in the generated
samplers will definitely be different, even if the object is
fundamentally similar or compatible.