Mark RMI Remote Object Config as a singleton.
Due to another misunderstanding on my part, the previous fix to make
the registry non-singleton was apparently unnecessary: JMeter actually
clones each test element for each thread, so that state variables
are *per-thread*. The ThreadLocal would therefore be unnecessary, in
that case.

However, because we need to actually share state across threads later
on, when we implement "global" mode, mark that the RMI Remote Object
Config test element not be cloned per thread, leaving it as a
singleton.